Move main widget out of index
